Checking homogeneity of variance

Returns information that can be used to determine if the homogeneity of variance, an assumption for ANCOVA, is met.

HomogeneityOfVariance(inputted.data, dependent.variable, independent.variable)

Arguments

inputted.data
A dataframe

dependent.variable
A string that specifies the column name of the column to use as the dependent variable. Column must be numeric.

independent.variable
A string that specifies the column name of the column to use as the independent variable. Column can be numeric or factor. If it's a factor, then it can only have two levels.

Value

A matrix with two rows. The first row specify what the values are in the second row. The second row: The first element is the formula used to evaluate Levene test. The next element is the p-value from the Levene test.

Details

To test homogeneity of regression slopes, the Levene test is used. If the p-value is significant, then this means the assumption of homogeneity of variance is not met.

Examples

dependent.col <- c(10.1, 11.3, 12.1, 13.7, 14.2, 1.6, 2.3, 3.2, 4.1, 5.3)
independent.col <- as.factor(c(1, 1, 1, 1, 1, 0, 0, 0, 0, 0))
covariate.one.col <- c(1, 2, 3, 4, 5, 1, 2, 3, 4, 5)
covariate.two.col <- as.factor(c(1, 0, 1, 0, 1, 0, 1, 0, 1, 0))

inputted.data <- data.frame(dependent.col, independent.col, covariate.one.col,
                            covariate.two.col)

results <- HomogeneityOfVariance(inputted.data, "dependent.col",
                                         "independent.col")

results

#>      [,1]                                                [,2]               
#> row1 "Levene Formula (test for homogeneity of variance)" "Levene P-value"   
#> row2 "dependent.col~independent.col"                     "0.717668447871424"